+++ bingads/v12/reporting/__init__.py
@@ -6,3 +6,5 @@ from .exceptions import *
 from .reporting_operation import *
 from .reporting_operation_status import *
 from .reporting_service_manager import *
+from .report_contract import *
+from .report_file_reader import *